[Feat] 허브 간 이동 경로 기능 및 레디스 연동

+++ core/src/main/java/com/sparta/core/dto/CompanyResponse.java
@@ -2,14 +2,15 @@ package com.sparta.core.dto;
 
 import com.sparta.core.entity.Company;
 import com.sparta.core.entity.CompanyTypeEnum;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CompanyResponse(
     String companyName,
     String address,
     CompanyTypeEnum companyType,
-    Long latitude,
-    Long longitude,
+    BigDecimal latitude,
+    BigDecimal longitude,
     UUID hubId,
     Long companyManagerId
 ) {
